+++ RELEASE_NOTES.md
@@ -93,6 +93,7 @@ If you have problems upgrading to `7.0`, first try upgrading to `6.9`. It has a 
 * `EmojiKeyboard` init parameters have been refactored. 
 * `InputCalloutContext` `.shared` has been removed.
 * `KeyboardAction` `.return` and `.newLine` have been replaced by `primary` variants.
+* `KeyboardAction` `.shift` `currentState` is renamed to `currentCasing`.
 * `KeyboardAction` `standardTextDocumentProxyAction` has been removed.
 * `KeyboardGesture` `.tap` has been replaced by `.release`.
 * `KeyboardInputViewController` has replaced `actionCalloutContext` and `inputCalloutContext` with a single `calloutContext`.